refactor: Encapsular archivos utilizados e implementar nuevos componentes de gestión de usuarios.
# ✅Cambios realizados
- Se encapsularon archivos en su carpeta con su tipo de funcionalidad para tener todo mas ordenado
- Se añadieron los componentes TableHeader y TableRow para la tabla de gestión de usuarios.
- Se implementó el componente FiltersTableUsers para filtrar los datos de los usuarios.
- Se crearon enlaces para las funciones de autenticación, registro y búsqueda de usuarios.
- Se introdujeron los enlaces useSidebar y useTheme para gestionar el estado y el tema de la barra lateral, respectivamente.
- Se actualizaron las funciones de servicio de usuario para crear, actualizar y eliminar usuarios.
- Se mejoró la gestión de usuarios con una interfaz de usuario y una gestión de estado optimizadas.
