+++ src/components/ui/sidebar/ContraerSidebar.tsx
@@ -1,4 +1,4 @@
-import { useSidebar } from "@/hooks/useSidebar";
+import { useSidebar } from "@/hooks/Sidebar/useSidebar";
 
 import { Icon } from "../Icon";
