+++ src/components/ui/sidebar/ImageSidebar.tsx
@@ -1,4 +1,4 @@
-import { useSidebar } from "@/hooks/useSidebar";
+import { useSidebar } from "@/hooks/Sidebar/useSidebar";
 import { v } from "@/styles/variables";
 
 export const ImageSidebar = () => {
