+++ src/components/ui/sidebar/SidebarItems.tsx
@@ -1,4 +1,4 @@
-import { useSidebar } from "@/hooks/useSidebar";
+import { useSidebar } from "@/hooks/Sidebar/useSidebar";
 
 import { NavLink } from "react-router";
