+++ src/components/ui/users/ModalEditarUsuario.tsx
@@ -1,4 +1,4 @@
-import { useUsers } from "@/hooks/useUsers";
+import { useUsers } from "@/hooks/Users/useUsers";
 
 import { useCallback } from "react";
