+++ src/pages/PrivateRoutes/Student.tsx
@@ -1,4 +1,4 @@
-import { useSidebar } from "@/hooks/useSidebar";
+import { useSidebar } from "@/hooks/Sidebar/useSidebar";
 
 export const Student = () => {
   const { isMobile, sidebarOpen } = useSidebar();
