+++ src/services/Users/userService.ts
@@ -1,4 +1,4 @@
-import { useAuthStore } from "@/store/authStore";
+import { useAuthStore } from "@/store/Auth/authStore";
 
 import { api } from "@config/api";
